+++ src/store/admin/state.js
@@ -2,4 +2,7 @@ export default {
   token: localStorage.getItem('auth-token') || '',
   isLoading: false,
   admin: {},
+  isDialogOpenned: false,
+  received: {},
+  isDisabled: false,
 }